restructure reactjs folder
