+++ src/components/Button/index.jsx
@@ -1,5 +1,3 @@
-import React from "react"
-
 export default function Button({type, children}) {
 	return (
 		<button className="gradient-button" type={type}>
