+++ src/components/Button/style.scss
@@ -1,0 +1,24 @@
+button.gradient-button {
+    background-image: $theme-orange;
+    padding: 10px 20px;
+    text-transform: uppercase;
+    
+    a{
+        color: white;
+        text-decoration: none;
+    }
+
+    background-size: 200% auto;
+    color: white;
+    box-shadow: 0 0 20px #eee;
+    border-radius: 10px;
+    display: block;
+    border: none;
+    cursor: pointer;
+    transition: 0.3s;
+    &:hover {
+        background-position: right center;
+        color: #fff;
+        text-decoration: none;
+    }
+}
