+++ src/components/NavBar/index.scss
