+++ src/components/TextInput/index.jsx
@@ -1,8 +1,3 @@
-import React from "react"
-
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { faUser } from "@fortawesome/free-solid-svg-icons"
-
 export default function TextInputForm({
 	type,
 	name,
