+++ src/components/Wrapper/index.jsx
@@ -1,5 +1,3 @@
-import React from "react"
-
 export default function Wrapper({ children }) {
 	return (
 		<div className="theme-root">
