+++ src/pages/Home/index.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import NavBar from "./NavBar"
+import NavBar from "../../components/NavBar"
 import "./styles.scss"
 
 export default function HomePage() {
