+++ src/pages/Home/styles.scss
@@ -1,10 +1,3 @@
-:root {
-	--dark: rgb(62, 92, 115);
-	--black: #123;
-	--white: rgb(255, 255, 255);
-	--menu: #e8e5e5;
-}
-
 /* Reset CSS */
 * {
 	padding: 0;
